Refactored and created modules
